+++ cli/ver.h
@@ -1,4 +1,0 @@
-#ifndef LIBTMAN_CLI_VER_H
-#define LIBTMAN_CLI_VER_H
-
-#endif